Split settings file into a subpackage
This will make it easier to continue expanding the schema to include the
schedule for executions.
